refactor: Replace ContractsIdentifier with ported impl
While the original port is nicer, I prefer to move over everything that
is already working with the existing source model. We should make this a
lot nicer and document this later on.
